Reshuffle file structure more
