+++ src/application/modules/composition/ui/Composition.tsx
@@ -10,11 +10,6 @@ import {
 
 import { InstallToolButton } from "@/application/modules/application-menu/ui/InstallToolButton.js";
 import { StartupScreen } from "@/application/modules/application-menu/ui/Startup.js";
-import LayerMouseControl, {
-  type DragState,
-} from "@/application/modules/canvas/ui/LayerMouseControl.js";
-import { MouseMode } from "@/application/modules/canvas/ui/MouseControl.js";
-import { ActionToolButton } from "@/application/services/actions/ActionToolButton";
 import { useFile } from "@/application/state/FileContext";
 import { useActionMap } from "@/application/state/hooks/useActionMap";
 import useEvent from "@/application/state/hooks/useEvent";
@@ -23,6 +18,11 @@ import {
   useScreenTranslation,
   useUpdateScreenTranslation,
 } from "@/application/state/ScreenTranslationContext";
+import { ActionToolButton } from "@/application/ui/ActionToolButton";
+import LayerMouseControl, {
+  type DragState,
+} from "@/application/ui/LayerMouseControl.js";
+import { MouseMode } from "@/application/ui/MouseControl.js";
 import { dragItem } from "@/domain/animation/file2/drag";
 import {
   findParentId,
