+++ src/application/modules/layers/ui/ShapeTree.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import { LayerTreeEnvironment } from "@/application/services/treeEnvironments/LayerTreeEnvironment.js";
+import { LayerTreeEnvironment } from "@/application/modules/layers/ui/LayerTreeEnvironment.js";
 import { useFile } from "@/application/state/FileContext";
 import { useToolAction } from "@/application/state/hooks/useToolAction";
 import {
