+++ src/domain/animation/file2/verifyFile.ts
@@ -2,6 +2,7 @@ import { type GeppettoImage } from "@/dtos/animation-file2.dto";
 
 import { newFile } from "./new";
 
+// TODO: Move to dto, use Zod to validate
 export const verifyFile = (file: unknown): file is GeppettoImage => {
   if (typeof file !== "object") return false;
   if (!file) return false;
