+++ src/domain/animation/readme.md
