+++ src/dtos/geppetto-file.ts
@@ -1,0 +1,34 @@
+import { verifyFile as verifyVersion1 } from "@/domain/animation/file1/verifyFile";
+import { convertFromV1 } from "@/domain/animation/file2/convert";
+import { verifyFile as verifyVersion2 } from "@/domain/animation/file2/verifyFile";
+import type { GeppettoImage } from "@/dtos/animation-file2.dto";
+
+export const loadGeppettoFile = async (
+  file: FileSystemFileHandle
+): Promise<[filename: string, image: GeppettoImage]> => {
+  const fileData = await file.getFile();
+  return new Promise((resolve) => {
+    const reader = new FileReader();
+
+    reader.addEventListener("load", () => {
+      const image = JSON.parse(reader.result as string);
+      if (verifyVersion1(image)) {
+        const version2 = convertFromV1(image);
+        resolve([file.name, version2]);
+      }
+      if (verifyVersion2(image)) {
+        resolve([file.name, image]);
+      }
+    });
+    reader.readAsText(fileData, "utf8");
+  });
+};
+
+export const saveGeppettoFile = async (
+  fileHandle: FileSystemFileHandle,
+  image: GeppettoImage
+): Promise<void> => {
+  const writable = await fileHandle.createWritable();
+  await writable.write(JSON.stringify(image, null, 2));
+  await writable.close();
+};
